Pivoting from a purely SQL database to a json hybrid
This should make it easier to handle the colorants as
to delete a colorant there will be no more need to store a list
of ids and check if the user wanted them removed. Instead the frontend
just send us a json string containing the json of the colorants that
the user has edited and stores that into that database.