Multiple bug fixes and refactors
* Removed all traces of trying to merge two structs its a hard problem
without any clear solutions. The main gripe is when a user clears
a field how do you know if that field was cleared on purpose or the user
just omitted it(in which case the old value was still needed).

All updates now need to be passed all values from the previous object
in order not to nullify those fields

* Fixing how Modals were loaded, previously they were watching a single
value and reloading based on that but this caused a bug where if the
data changed but the user opened the same modal twice it wouldn't load
the latest data. Now Modals are loaded on click