Adding better search functionality
Used bleve to achieve in-memory search. This has the drawback
of hindering scaleability, but gives overall better search
results than the rudimentary filtering we were doing before.

The biggest improvement is the ability to search for things
that don't appear right in the table. So its now possible
to search for things like notes.