Move unlink/link to datastore functions
Previously we were unlinking and linking formulas to jobs after each
update/add call in the handler function. This is prone to the database
becoming out of sync as one of the datastore functions could succeed
while the others fail even though they depend on each other.

Moved the linking and unlinking of formulas/jobs to the datastore
instead to the datastore so they're all under a single transaction

Also fixed a problem with createFormulasModal not persisting jobs
correctly.

Fixes #44