Fix search to match on dashed queries
Previously queries were unsanitized for bleve search operators. This
caused users to attempt a search for a resource like "hello-world" and
instead of matching on a document with "hello-world" instead the search
would return the intersection of results where "hello" was present and
"world" was not (treating the '-' operator as a "must not exist").

Fixes #50